pyproject.toml
[project]
name = "fly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "albumentations>=1.3.0",
    "ultralytics>=8.3.228",
    "matplotlib>=3.3.0",
    "opencv-python>=4.6.0",
    "pillow>=7.1.2",
    "numpy<2.0",
    "pyyaml>=5.3.1",
    "scipy>=1.4.1",
    "tqdm>=4.64.0", # progress bars
    "psutil", # system utilization
    "py-cpuinfo", # display CPU info
    "thop>=0.1.1", # FLOPs computation
    "pandas>=1.1.4",
    "seaborn>=0.11.0", # plotting
    "pyqt6>=6.4.2",
    "pyqt6-tools>=6.4.2.3.3",
    "torch==2.8.0",
    "torchvision>=0.9.0",
]



#[build-system]
#requires = ["setuptools>=68", "wheel>=0.43"]
#build-backend = "setuptools.build_meta"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
